+++ server/component/style/entryList/EntryListModel.php
@@ -38,9 +38,24 @@ class EntryListModel extends StyleModel
      */
     private $own_entries_only;
 
-    public function __construct($services, $id)
+    /**
+     * The constructor
+     *
+     * @param array $services
+     *  An associative array holding the different available services. See the
+     *  class definition basepage for a list of all services.
+     * @param int $id
+     *  The id of the section with the conditional container style.
+     * @param array $params
+     *  The list of get parameters to propagate.
+     * @param number $id_page
+     *  The id of the parent page
+     * @param array $entry_record
+     *  An array that contains the entry record information.
+     */
+    public function __construct($services, $id, $params, $id_page, $entry_record)
     {
-        parent::__construct($services, $id);
+        parent::__construct($services, $id, $params, $id_page, $entry_record);
         $this->init_properties();
     }