+++ server/component/style/tableRow/TableRowView.php
@@ -1,0 +1,40 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<?php
+require_once __DIR__ . "/../StyleView.php";
+
+/**
+ * The view class of the Table Row style component.
+ * A table style is a container that allows to wrap content into table.
+ */
+class TableRowView extends StyleView
+{
+
+    /* Constructors ***********************************************************/
+
+    /**
+     * The constructor.
+     *
+     * @param object $model
+     *  The model instance of the component.
+     */
+    public function __construct($model)
+    {
+        parent::__construct($model);
+    }
+
+    /* Public Methods *********************************************************/
+
+    /**
+     * Render the style view.
+     */
+    public function output_content()
+    {
+        require __DIR__ . "/tpl_tableRow.php";
+    }
+
+}
+?>